+++ code/src/App.js
@@ -1,31 +1,84 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 
 import { NewMessageForm } from './components/NewMessageForm'
 import { Messages} from './components/Messages'
 
+import { API_URL, LIKE_API } from './reusables/urls'
+
 export const App = () => {
 
-  const [newMessage, setNewMessage] = useState('')
   const [messageList, setMessageList] = useState([])
+  const [newMessage, setNewMessage] = useState('')
 
-  const handleMessageList = (newPost) => {
-    setMessageList([...messageList, newPost])
-  }
+  useEffect(() => {
+    fetchList()
+    }, []);//OK
+
+    const fetchList = () => {
+      fetch(API_URL)
+        .then(res => res.json())
+        .then(messages => setMessageList(messages))
+        .catch(error => console.error(error))
+    }//OK
+
+    const handleNewMessageChange = (e) => {
+      setNewMessage(e.target.value)
+    }//OK
+
+    const handleSubmit = (e) => {
+      e.preventDefault()
+    
+      const options = {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ message: newMessage })
+      };
+  
+      fetch(API_URL, options)
+        .then(res => res.json())
+        .then(() => fetchList())
+        .catch(error => console.error(error))
+        setNewMessage('')
+    } //OK
+
+    const options = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+    };
+
+    const handleIncreaseLikes = (id) => {
+
+      fetch(LIKE_API(id), options)
+      .then(res => res.json())
+      .then(newLike => {
+        const updatedMessageList = messageList.map(message => {
+          if (message._id === newLike._id) {
+            message.hearts += 1;
+          }
+          return message
+          }) 
+          setMessageList(updatedMessageList)
+      })
+      .catch(error => console.error(error))
+    } //OK
 
 
   return (
     <div className='main'>
+
       <NewMessageForm
         newMessage={newMessage}
-        setNewMessage={setNewMessage}
-        messageList={messageList}
-        setMessageList={setMessageList}
-        handleMessageList={handleMessageList}
+        onSubmit={handleSubmit}
+        onNewMessageChange={handleNewMessageChange}
       />
     
       <Messages
         messageList={messageList}
-        setMessageList={setMessageList}
+        handleIncreaseLikes={handleIncreaseLikes}
       />
     </div>
