+++ code/src/components/Messages.js
@@ -1,49 +1,8 @@
-import React, {useEffect} from 'react'
-import moment from 'moment'
+import React from 'react'
 
-import { API_URL, LIKE_API } from '../reusables/urls'
+import { MessageElement } from './MessageElement'
 
-export const Messages = ({ messageList, setMessageList }) => {
-
-    useEffect(() => {
-      //fetch for list of messages    
-      const fetchList = (url) => {
-        fetch(url)
-          .then(res => res.json())
-          .then(messages => setMessageList(messages))
-          .catch(error => console.error(error))
-      }
-        fetchList(API_URL)
-      }, [messageList]);
-
-
-
-//fetch for increasing likes
-
-      const options = {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-      };
-
-      const increaseLikes = (id) => {
-
-        fetch(LIKE_API(id), options)
-        .then(res => res.json())
-        .then(newLike => {
-          const updatedMessageList = messageList.map(message => {
-            if (message._id === newLike._id) {
-              message.hearts += 1;
-            }
-            return message
-            }) 
-            setMessageList(updatedMessageList)
-
-        })
-        .catch(error => console.error(error))
-      }
-
+export const Messages = ({ messageList, handleIncreaseLikes}) => {
 
 
     return (
@@ -52,30 +11,11 @@ export const Messages = ({ messageList, setMessageList }) => {
             <div 
               key={message._id}
               className='messageContent'>
-              <h4 className='messageText'>
-                {message.message}
-              </h4>
-              <div className='heartAndTime'>
-                <div className='heartThings'>
-                  <button
-                    className='heartButton'
-                    style={{ background: message.hearts >= 1 ? '#ffadad' : '#eaeaea'}}
-                    onClick={() => increaseLikes(message._id)}>
-                    <span
-                      className='heart'
-                      role='img'
-                      aria-label='heart emoji'>
-                        &#128151;
-                    </span>
-                  </button>
-                  <p className='nrOfLikes'>
-                    x {message.hearts}
-                  </p>
-                </div>
-                <p className='time'>
-                  {moment(message.createdAt).fromNow()}
-                </p>
-              </div>
+                <MessageElement
+                  key={message._id}
+                  message={message}
+                  onIncreaseLikes={handleIncreaseLikes}
+                />
             </div>
           ))}
         </div>
